pyproject.toml
[tool.poetry]
name = "claude-code-openai-wrapper"
version = "1.0.0"
description = "OpenAI API-compatible wrapper for Claude Code"
authors = ["Richard Atkinson <richardatk01@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
sse-starlette = "^2.1.3"
python-multipart = "^0.0.18"
claude-code-sdk = "0.0.22"
slowapi = "^0.1.9"
tavily-python = "^0.7.12"
psutil = "^6.1.0"
pytz = ">=2024.1"

# AWS Bedrock support (DSGVO-compliant backend)
boto3 = "^1.35.0"

# Privacy/DSGVO compliance (Presidio-based PII anonymization)
presidio-analyzer = {version = "^2.2.0", optional = true}
presidio-anonymizer = {version = "^2.2.0", optional = true}

[tool.poetry.extras]
privacy = ["presidio-analyzer", "presidio-anonymizer"]

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
requests = "^2.32.0"
openai = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.poetry.scripts]
claude-wrapper = "src.main:run_server"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]  # KRITISCH: Fügt project root zu PYTHONPATH hinzu
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"